requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
pyyaml==6.0.1
playwright==1.44.0
